Created week selection component
